+++ trixnity-messenger/src/commonMain/kotlin/de/connect2x/trixnity/messenger/util/MatrixReferences.kt
@@ -3,10 +3,9 @@ package de.connect2x.trixnity.messenger.util
 import de.connect2x.trixnity.messenger.util.html.HtmlNode
 import de.connect2x.trixnity.messenger.util.html.MatrixMentionVisitor
 import net.folivo.trixnity.core.util.Reference
-import net.folivo.trixnity.core.util.References
 
 internal object MatrixReferences {
-    fun findInText(text: String): Map<IntRange, Reference> = References.findReferences(text)
+    fun findInText(text: String): Map<IntRange, Reference> = Reference.findReferences(text)
 
     fun findInHtml(content: HtmlNode): Map<String, Reference> =
         MatrixMentionVisitor().process(content)